backend/requirements.txt
flask
flask-cors
openai
moviepy
networkx
numpy
pandas
torch
torchaudio
torchvision
transformers
accelerate
pillow
nano-vectordb
timm
einops